% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_co2_footprint_descriptive_analysis_per_benchmark.R
\name{calculate_co2_footprint_descriptive_analysis_per_benchmark}
\alias{calculate_co2_footprint_descriptive_analysis_per_benchmark}
\title{Calculate the descriptive analysis of co2_footprint per benchmark}
\usage{
calculate_co2_footprint_descriptive_analysis_per_benchmark(
  emission_product,
  benchmark
)
}
\arguments{
\item{emission_product}{A dataframe. Product-level output of Emissions profile}

\item{benchmark}{A string. Benchmarks of Emissions profile}
}
\value{
A dataframe
}
\description{
This function calculates the descriptive analysis of co2_footprint per
benchmark using product-level output of Emissions profile indicator.
}
\examples{
library(dplyr)
# styler: off
emission_product_example <- tribble(
          ~benchmark, ~tilt_subsector, ~co2_footprint,   ~ep_product,
    "tilt_subsector",     "chemicals",              1,       "water",
    "tilt_subsector",     "chemicals",              4,  "television",
    "tilt_subsector",     "chemicals",              9,      "tables",
    "tilt_subsector",     "chemicals",              8, "electricity",
    "tilt_subsector",     "chemicals",       NA_real_,        "meat",
    "tilt_subsector",  "iron & steel",              1,      "tables",
    "tilt_subsector",  "iron & steel",              2,       "water",
    "tilt_subsector",  "iron & steel",              4,        "tent"
  )
# styler: on

emission_product_example

# Calculates the descriptive analysis for benchmark "tilt_subsector"
calculate_co2_footprint_descriptive_analysis_per_benchmark(
  emission_product_example,
  "tilt_subsector"
)
}
