% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_trs_descriptive_analysis_per_scenario_year.R
\name{calculate_trs_descriptive_analysis_per_scenario_year}
\alias{calculate_trs_descriptive_analysis_per_scenario_year}
\title{Calculate the descriptive analysis of transition risk scores for all
\code{tilt_subsectors} per scenario-year benchmark}
\usage{
calculate_trs_descriptive_analysis_per_scenario_year(
  transition_risk_product,
  scenario_year_benchmark
)
}
\arguments{
\item{transition_risk_product}{A dataframe. Product-level output of
Transition risk profile}

\item{scenario_year_benchmark}{A string. Benchmarks of Sector profile}
}
\value{
A dataframe
}
\description{
This function calculates the descriptive analysis of transition risk scores
for all \code{tilt_subsectors} per scenario-year benchmark using product-level
output of Transition risk profile indicator.
}
\examples{
library(dplyr)
transition_risk_product_example <- product_transition_risk |>
  select(c("tilt_subsector", "scenario", "year", "transition_risk_score")) |>
  distinct()
transition_risk_product_example

# Calculates the descriptive analysis for benchmark "1.5C RPS_2030"
calculate_trs_descriptive_analysis_per_scenario_year(
  transition_risk_product_example,
  "1.5C RPS_2030"
)
}
