% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_emission}
\alias{company_emission}
\title{Company-level data of Emissions profile}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1728 rows and 15 columns.
}
\usage{
company_emission
}
\description{
Company-level data of Emissions profile
}
\keyword{datasets}
