% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_sector}
\alias{company_sector}
\title{Company-level data of Sector profile}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1152 rows and 13 columns.
}
\usage{
company_sector
}
\description{
Company-level data of Sector profile
}
\keyword{datasets}
