% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_transition_risk}
\alias{company_transition_risk}
\title{Company-level data of Transition risk profile}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1728 rows and 38 columns.
}
\usage{
company_transition_risk
}
\description{
Company-level data of Transition risk profile
}
\keyword{datasets}
