% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_emission_product_co2_des_analysis}
\alias{example_emission_product_co2_des_analysis}
\alias{example_transition_risk_product_trs_des_analysis}
\alias{example_transition_risk_emission_ep_product_des_analysis}
\alias{example_transition_risk_sector_ep_product_des_analysis}
\alias{example_transition_risk_ep_product_des_analysis}
\alias{example_transition_risk_product_emission_cov}
\alias{example_transition_risk_product_sector_cov}
\alias{example_transition_risk_product_transition_risk_cov}
\alias{example_emission_product_best_case_worst_case}
\alias{example_sector_product_best_case_worst_case}
\alias{example_transition_risk_company_emission_avg_best_case_worst_case}
\alias{example_transition_risk_company_sector_avg_best_case_worst_case}
\alias{example_transition_risk_product_best_case_worst_case}
\alias{example_transition_risk_company_transition_risk_avg_best_case_worst_case}
\title{Example raw datasets}
\usage{
example_emission_product_co2_des_analysis()

example_transition_risk_product_trs_des_analysis()

example_transition_risk_emission_ep_product_des_analysis()

example_transition_risk_sector_ep_product_des_analysis()

example_transition_risk_ep_product_des_analysis()

example_transition_risk_product_emission_cov()

example_transition_risk_product_sector_cov()

example_transition_risk_product_transition_risk_cov()

example_emission_product_best_case_worst_case()

example_sector_product_best_case_worst_case()

example_transition_risk_company_emission_avg_best_case_worst_case()

example_transition_risk_company_sector_avg_best_case_worst_case()

example_transition_risk_product_best_case_worst_case()

example_transition_risk_company_transition_risk_avg_best_case_worst_case()
}
\value{
A dataframe.
}
\description{
These datasets shows a state of the data prior to becoming an input for this
package.
}
\examples{
example_emission_product_co2_des_analysis()
example_transition_risk_product_trs_des_analysis()
example_transition_risk_product_emission_cov()
example_transition_risk_product_sector_cov()
example_transition_risk_product_transition_risk_cov()
example_emission_product_best_case_worst_case()
example_transition_risk_company_emission_avg_best_case_worst_case()
example_transition_risk_company_sector_avg_best_case_worst_case()
example_transition_risk_product_best_case_worst_case()
example_transition_risk_company_transition_risk_avg_best_case_worst_case()
example_transition_risk_emission_ep_product_des_analysis()
example_transition_risk_sector_ep_product_des_analysis()
example_transition_risk_ep_product_des_analysis()
example_sector_product_best_case_worst_case()
}
\keyword{internal}
