% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_emission}
\alias{product_emission}
\title{Product-level data of Emissions profile}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 456 rows and 33 columns.
}
\usage{
product_emission
}
\description{
Product-level data of Emissions profile
}
\keyword{datasets}
