% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_sector}
\alias{product_sector}
\title{Product-level data of Sector profile}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 304 rows and 28 columns.
}
\usage{
product_sector
}
\description{
Product-level data of Sector profile
}
\keyword{datasets}
