% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_transition_risk}
\alias{product_transition_risk}
\title{Product-level data of Transition risk profile}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1824 rows and 42 columns.
}
\usage{
product_transition_risk
}
\description{
Product-level data of Transition risk profile
}
\keyword{datasets}
