% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/show_reduction_targets_for_tilt_subsectors_per_scenario_year.R
\name{show_reduction_targets_for_tilt_subsectors_per_grouping_sector}
\alias{show_reduction_targets_for_tilt_subsectors_per_grouping_sector}
\title{Show reduction targets for all \code{tilt_subsectors} per grouping_sector}
\usage{
show_reduction_targets_for_tilt_subsectors_per_grouping_sector(sector_product)
}
\arguments{
\item{sector_product}{A dataframe. Product-level output of Sector profile}
}
\value{
A dataframe
}
\description{
This function shows reduction targets for all \code{tilt_subsectors} per
grouping_sector using product-level output of Sector profile indicator.
}
\examples{
library(dplyr)
sector_product_example <- product_sector |>
  select(c("tilt_subsector", "scenario", "year", "reduction_targets")) |>
  distinct()
sector_product_example

show_reduction_targets_for_tilt_subsectors_per_grouping_sector(
  sector_product_example
)
}
