% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltDataAnalysis-package.R
\docType{package}
\name{tiltDataAnalysis-package}
\alias{tiltDataAnalysis}
\alias{tiltDataAnalysis-package}
\title{tiltDataAnalysis: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/tiltDataAnalysis}
  \item \url{https://2degreesinvesting.github.io/tiltDataAnalysis/}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltDataAnalysis/issues}
}

}
\author{
\strong{Maintainer}: Kalash Singhal \email{kalash@2degrees-investing.org}

Authors:
\itemize{
  \item Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})
  \item Anne Schoenauer \email{anne.schoenauer@2degrees-investing.org} (\href{https://orcid.org/0000-0002-4576-8799}{ORCID})
  \item Tilman Trompke \email{tilman@2degrees-investing.org} (\href{https://orcid.org/0009-0003-3351-5131}{ORCID})
}

}
\keyword{internal}
