% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_emission_product_co2_des_analysis}
\alias{example_emission_product_co2_des_analysis}
\alias{example_transition_risk_product_trs_des_analysis}
\alias{example_emission_ep_product_des_analysis}
\alias{example_sector_ep_product_des_analysis}
\title{Example raw datasets}
\usage{
example_emission_product_co2_des_analysis()

example_transition_risk_product_trs_des_analysis()

example_emission_ep_product_des_analysis()

example_sector_ep_product_des_analysis()
}
\value{
A dataframe.
}
\description{
These datasets shows a state of the data prior to becoming an input for this
package.
}
\examples{
example_emission_product_co2_des_analysis()
example_transition_risk_product_trs_des_analysis()
example_emission_ep_product_des_analysis()
example_sector_ep_product_des_analysis()
}
\keyword{internal}
